pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
name = "spotify-podcast-rss"
authors = [
    {name = "Robert Huselius", email = "robert@huseli.us"},
]
description = "RSS for Spotify podcasts"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
dependencies = [
    "feedgen",
    "python-dateutil",
    "spotipy",
]
dynamic = ["version"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Sound/Audio",
]

[project.urls]
Repository = "https://github.com/Eboreg/spotify-podcast-rss"

[project.optional-dependencies]
dev = [
    "ipdb",
    "ipython",
    "isort",
    "pylint",
    "types-python-dateutil",
]

[project.scripts]
spr = "spr.cli:main"

[tool.setuptools.dynamic]
version = {attr = "spr.__version__"}

[tool.pylint.main]
disable = [
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "broad-exception-caught",
    "too-many-branches",
]
ignore = [".venv"]
max-line-length = 119

[tool.isort]
combine_as_imports = true
default_section = "THIRDPARTY"
include_trailing_comma = true
line_length = 79
multi_line_output = 5
known_self = "spr"
lines_after_imports = 2
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "SELF",
    "LOCALFOLDER",
]
skip = [".venv"]

[tool.mypy]
follow_imports = "skip"
show_error_codes = true
ignore_missing_imports = true
files = "spr/."
check_untyped_defs = true
